+++ CHANGELOG.md
@@ -28,6 +28,8 @@ For more information on SemVer, please visit http://semver.org/.
 * Added `wpdk_flush_cache_third_parties_plugins` action in order to flush third parties plugins.
 * Added `WPDKUIControlType::inputTypeWithClass()` helper static method in order to retrive the HTML control type string.
 * Fixed dynamic table css styles.
+* Fixed potential Javascript warning in localize script on jQuery
+* Minor stability fixes
 
 
 ## Version 1.7.2